usm: kafka: Change topic name size to u8
The topic name has an upper limit of 255 by the protocol. Also, we have a lower liimt of 80
characters. Thus, using u16 for the topic name size is a waste.

+++ pkg/network/protocols/kafka/statkeeper_test.go
@@ -24,7 +24,7 @@ func BenchmarkStatKeeperSameTX(b *testing.B) {
 
 	tx := new(KafkaTransaction)
 	copy(tx.Topic_name[:], topicName)
-	tx.Topic_name_size = uint16(topicNameSize)
+	tx.Topic_name_size = uint8(topicNameSize)
 
 	b.ReportAllocs()
 	b.ResetTimer()
